+++ packages/sdks/src/blocks/image/image.lite.tsx
@@ -1,4 +1,10 @@
-import { Show, useMetadata, useStore } from '@builder.io/mitosis';
+import {
+  onMount,
+  Show,
+  useMetadata,
+  useStore,
+  useTarget,
+} from '@builder.io/mitosis';
 import type { JSX } from '@builder.io/mitosis/jsx-runtime';
 import { getSrcSet } from './image.helpers.js';
 import type { ImageProps } from './image.types.js';
@@ -61,6 +67,21 @@ export default function Image(props: ImageProps) {
       const out = props.aspectRatio ? aspectRatioStyles : undefined;
       return out;
     },
+  });
+
+  onMount(() => {
+    useTarget({
+      angular: () => {
+        /** this is a hack to include the input in angular */
+        const _ = {
+          a: props.lockAspectRatio,
+          b: props.width,
+          c: props.height,
+          d: props.lazy,
+          e: props.attributes,
+        };
+      },
+    });
   });
   return (
     <>